Rename `lib/utils` to `libs/utils` and update imports
This commit updates the utility functions by renaming `@/lib/utils` to `@/libs/utils`. Consequently, all relevant import paths in the project have been adjusted accordingly to maintain consistency and avoid import errors.
