+++ app/_components/ui/input-otp.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import { OTPInput, OTPInputContext } from 'input-otp';
 import { Dot } from 'lucide-react';
 
-import { cn } from '@/lib/utils';
+import { cn } from '@/libs/utils';
 
 const InputOTP = React.forwardRef<
   React.ElementRef<typeof OTPInput>,
